Expanded trace_gui, the application no longer shuts down when a the Keithley object fails to initiate due to a lack of connection. An error warning pops up.
Further more the trace_gui only takes data prerecorded manualy from the source meter. The idea behind this is to permit higher temporal measurement resolution.

The GUI shows two buttons acompanied by edit lines. In the first one the name of the communications port should be inputted by the user. The second one is for the desired filename. Next to the desired filename field is a dropdown menu for the desired file extension.
Pressing connect, initialises serial communication with keithley and instantiates a Keithley object from keithley_control.
Pressing trace, extracts the keithley trace buffer data and saves it under the desired filename.